added predicate function
